requirements.txt
fastapi
uvicorn
tensorflow
numpy
pandas
seaborn
matplotlib
requests
python-dotenv